+++ packages/components/src/notification/Notification.js
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import ReactDOM from 'react-dom';
-import { createContainer, execute, isNil, isString, nonNil, validate } from '../Utils';
+import { createContainer, execute, isNil, isString, nonNil, renderDom, validate } from '../Utils';
 import Alert from '../alert';
 import { to, useTransition } from 'react-spring';
 
@@ -210,7 +210,7 @@ const send = (type, cfg) => {
     let containerObj = createContainer('notify-' + msg.position);
 
     //pass position into todo
-    ReactDOM.render(
+    renderDom(
       <Notification
         position={msg.position}
         msgStore={proxy}